ADDED: FrmVentas funcionando, Login funcionando

+++ DataLayer/Models/DetalleEntrada.cs
@@ -5,7 +5,7 @@
         public int id_detalle_entrada { get; set; }
         public int id_compra { get; set; }
         public int id_producto { get; set; }
-        public int cantidad { get; set; }
+        public decimal cantidad { get; set; }
         public decimal precio_unidad { get; set; }
     }
