+++ DataLayer/Models/ProductoAlmacen.cs
@@ -5,7 +5,7 @@
         public int id_producto_almacen { get; set; }
         public int id_almacen { get; set; }
         public int id_producto { get; set; }
-        public int stock_actual { get; set; }
+        public decimal stock_actual { get; set; }
     }
 
     public class ProductoAlmacen : ViProductoAlmacen
